backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
pydantic
pydantic-settings
python-dotenv
python-multipart
python-jose
bcrypt
reportlab
